+++ testing/test_terminal.py
@@ -370,6 +370,31 @@ class TestFixtureReporting:
             "*1 failed*1 error*",
          ])
 
+    def test_setup_teardown_output_and_test_failure(self, testdir):
+        """ Test for issue #442 """
+        testdir.makepyfile("""
+            def setup_function(function):
+                print ("setup func")
+
+            def test_fail():
+                assert 0, "failingfunc"
+
+            def teardown_function(function):
+                print ("teardown func")
+        """)
+        result = testdir.runpytest()
+        result.stdout.fnmatch_lines([
+            "*test_fail*",
+            "*def test_fail():",
+            "*failingfunc*",
+            "*Captured stdout setup*",
+            "*setup func*",
+            "*Captured stdout teardown*",
+            "*teardown func*",
+
+            "*1 failed*",
+         ])
+
 class TestTerminalFunctional:
     def test_deselected(self, testdir):
         testpath = testdir.makepyfile("""